feat: schema enhancement step 1: Register SchemaExecutor in Execution Pipeline
