+++ notes/Github上的PR处理流程.md
@@ -1,0 +1,139 @@
+非常好的问题！这正是一个项目从“计划”走向“实践”的关键一步。你的理解非常接近了，只需要做一些微调，就能形成一套非常专业和安全的工作流程。
+
+我们来一步步拆解，并给你一套清晰的最佳实践。
+
+### 核心概念：PR 不是分支，而是“合并请求”
+
+让我们先理清这两个概念的关系，这很重要：
+
+*   **分支 (Branch)**：是你的代码库的一个**独立副本**。你可以把它想象成你正在写一本书 (`main` 分支)，为了写一个新的章节，你把书复印一份 (`feature` 分支)，然后在复印件上修改，完全不影响原稿。
+*   **拉取请求 (Pull Request, PR)**：它**不是**一个分支。它是一个**动作**，一个**通知**。当你完成了新章节的草稿（在 `feature` 分支上提交了代码），你发起一个 PR，就相当于给主编（在这里是你自己）发了一封邮件说：“嘿，我写好了‘CLI 骨架’这一章，请审阅一下。如果没问题，请把它合并到正式的书稿 (`develop` 或 `main` 分支) 里去吧。”
+
+所以，**PR 是用来讨论和审查一个分支上的代码，并请求将其合并到另一个分支的工具**。
+
+---
+
+### 你的问题的直接回答
+
+> **“在真正开始 project 之前 ，我是否需要先建立一个分支？”**
+
+**是的，但不完全是。** 在你写任何一行代码之前，你应该确保你的 `main` 分支是干净的、初始化的。然后，对于你项目板上的**每一个任务 (Issue)**，你都应该创建一个**新的、专门的**分支。
+
+**永远不要直接在 `main` 分支上写代码！** 这是第一金科玉律。
+
+> **“那么建立 PR 是否又是另一个分支？”**
+
+不是。建立 PR 是把你已经存在的 `feature` 分支提交给另一个分支（比如 `main` 或 `develop`）进行评审和合并。一个 PR 总是关联着一个**源分支**（你的功能分支）和一个**目标分支**。
+
+> **“所谓合并 PR 就是把 PR 所在的分支合并到先前的分支上”**
+
+**完全正确！** 当你在 GitHub 上点击 “Merge Pull Request” 按钮时，GitHub 就会在后台执行一个 `git merge` 命令，把你的功能分支（源分支）上的所有更改合并到目标分支中去。
+
+> **“只有最终 release 我才需要把所有分支合并到 main”**
+
+这个想法非常棒，说明你已经有了版本发布和保持 `main` 分支稳定的意识。这正是专业工作流的核心思想。
+
+为了让这个想法更具操作性，我们引入一个中间层，这就是大名鼎鼎的 **Git Flow** 思想的简化版，非常适合单人或小团队。
+
+---
+
+### 推荐给你的工作流程 (Solo Dev 最佳实践)
+
+忘掉一次性把所有分支都合到 `main` 的想法，那样会很混乱。采用下面这个更清晰、更安全的流程：
+
+#### 第 1 步：建立两个“永久”分支
+
+1.  **`main` 分支**：
+    *   **用途**：神圣不可侵犯！它只用来存放**已发布的、稳定的、可随时部署的**版本。
+    *   **规则**：你永远不直接向 `main` 提交代码。唯一的例外是通过发布流程从 `develop` 分支合并。你甚至可以在 GitHub 设置中将 `main` 分支保护起来，禁止直接推送。
+
+2.  **`develop` 分支** (或叫 `dev`)：
+    *   **用途**：这是你的**开发主线**。它集成了所有已完成的功能，代表了下一个版本“正在开发中”的最新状态。这个分支应该是大部分时间里功能最全，但可能不完全稳定的地方。
+    *   **操作**：`git checkout -b develop` 从 `main` 创建一次，然后 `git push -u origin develop` 推送到远程。之后它就一直存在了。
+
+#### 第 2 步：针对每个任务的工作流
+
+现在，假设你要开始做项目板上的任务 `Feature: CLI command schema – command skeleton`。
+
+1.  **准备**：确保你的本地 `develop` 分支是最新状态。
+    ```bash
+    git checkout develop
+    git pull origin develop
+    ```
+
+2.  **创建功能分支**：从 `develop` 分支上创建一个新的、描述清晰的功能分支。
+    ```bash
+    # 格式：类型/任务简称
+    git checkout -b feature/cli-skeleton
+    ```
+
+3.  **编码**：在这个 `feature/cli-skeleton` 分支上安心地写代码、修改、测试。频繁地提交 (commit) 是个好习惯。
+    ```bash
+    git add .
+    git commit -m "feat: add basic structure for schema command"
+    # ...继续工作和提交...
+    ```
+
+4.  **推送分支**：当你觉得这个功能基本完成，需要评审或备份时，就把它推送到远程 GitHub 仓库。
+    ```bash
+    git push -u origin feature/cli-skeleton
+    ```
+
+5.  **创建 PR**：现在去你的 GitHub 仓库页面。GitHub 会自动检测到你推送了一个新分支，并提示你创建一个 Pull Request。
+    *   **源分支 (from)**：`feature/cli-skeleton`
+    *   **目标分支 (to)**：`develop`  <-- **注意：是合并到 `develop`，不是 `main`！**
+    *   在 PR 的描述里，链接到你正在做的那个 Issue (例如，写上 `Resolves #2`，#2 是 Issue 的编号)。
+
+6.  **自我审查与合并**：
+    *   即使只有你一个人，也要**认真审查你自己的 PR**。在 “Files changed” 标签页里检查你的每一行改动，这是发现低级错误和改进代码的绝佳机会。
+    *   确认无误后，点击 “Merge Pull Request”。
+    *   合并后，可以选择删除远程的 `feature/cli-skeleton` 分支，因为它已经完成了使命。
+
+7.  **循环**：对于项目板上的下一个任务，重复步骤 1-6。
+
+#### 第 3 步：发布版本 (Release)
+
+当你完成了里程碑 (`v0.1.0`) 中的所有任务，并且它们都已经被合并到了 `develop` 分支，你的 `develop` 分支现在就包含了所有新功能，并且经过了测试。
+
+这时，就到了**发布**的时刻！
+
+1.  **创建发布 PR**：创建一个**新的 PR**，这次是从 `develop` 合并到 `main`。
+    *   **源分支 (from)**：`develop`
+    *   **目标分支 (to)**：`main`
+    *   这个 PR 的标题可以命名为 `Release v0.1.0`。
+
+2.  **最终检查**：这个 PR 包含了你这段时间以来的所有心血。做最后的检查。
+
+3.  **合并到 `main`**：合并这个 PR。现在，你的 `main` 分支就更新到了最新的 `v0.1.0` 版本。
+
+4.  **打上标签 (Tag)**：在 `main` 分支上创建一个 Git 标签，以标记这个发布点。
+    ```bash
+    git checkout main
+    git pull origin main
+    git tag -a v0.1.0 -m "Release version 0.1.0"
+    git push origin v0.1.0
+    ```
+
+### 总结与可视化
+
+你的分支结构看起来会是这样：
+
+```
+main      (稳定版)  <--- [仅在发布时合并] --- develop (开发版)
+                                                ^
+                                                | [PR合并]
+                                                |
+feature/task-A --- (完成) --------------------/
+                                                ^
+                                                | [PR合并]
+                                                |
+feature/task-B --- (完成) --------------------/
+```
+
+这个流程的好处是：
+*   **安全**：`main` 分支永远是干净的。
+*   **清晰**：每个功能都在独立的分支开发，互不干扰。
+*   **有序**：所有完成的功能汇集到 `develop`，便于集成测试和版本准备。
+*   **可追溯**：每个版本发布都有一个清晰的从 `develop`到 `main` 的 PR 和一个 Git 标签。
+
+这套流程初看步骤多，但一旦习惯，你会发现它极大地提升了你的开发质量和信心。祝你编码愉快！
