+++ notes/Github中项目管理流程.md
@@ -1,0 +1,202 @@
+#Github项目管理流程
+
+### 核心概念：PR 不是分支，而是“合并请求”
+
+让我们先理清这两个概念的关系，这很重要：
+
+*   **分支 (Branch)**：是你的代码库的一个**独立副本**。你可以把它想象成你正在写一本书 (`main` 分支)，为了写一个新的章节，你把书复印一份 (`feature` 分支)，然后在复印件上修改，完全不影响原稿。
+*   **拉取请求 (Pull Request, PR)**：它**不是**一个分支。它是一个**动作**，一个**通知**。当你完成了新章节的草稿（在 `feature` 分支上提交了代码），你发起一个 PR，就相当于给主编（在这里是你自己）发了一封邮件说：“嘿，我写好了‘CLI 骨架’这一章，请审阅一下。如果没问题，请把它合并到正式的书稿 (`develop` 或 `main` 分支) 里去吧。”
+
+所以，**PR 是用来讨论和审查一个分支上的代码，并请求将其合并到另一个分支的工具**。
+
+---
+
+### 你的问题的直接回答
+
+> **“在真正开始 project 之前 ，我是否需要先建立一个分支？”**
+
+**是的，但不完全是。** 在你写任何一行代码之前，你应该确保你的 `main` 分支是干净的、初始化的。然后，对于你项目板上的**每一个任务 (Issue)**，你都应该创建一个**新的、专门的**分支。
+
+**永远不要直接在 `main` 分支上写代码！** 这是第一金科玉律。
+
+> **“那么建立 PR 是否又是另一个分支？”**
+
+不是。建立 PR 是把你已经存在的 `feature` 分支提交给另一个分支（比如 `main` 或 `develop`）进行评审和合并。一个 PR 总是关联着一个**源分支**（你的功能分支）和一个**目标分支**。
+
+> **“所谓合并 PR 就是把 PR 所在的分支合并到先前的分支上”**
+
+**完全正确！** 当你在 GitHub 上点击 “Merge Pull Request” 按钮时，GitHub 就会在后台执行一个 `git merge` 命令，把你的功能分支（源分支）上的所有更改合并到目标分支中去。
+
+> **“只有最终 release 我才需要把所有分支合并到 main”**
+
+这个想法非常棒，说明你已经有了版本发布和保持 `main` 分支稳定的意识。这正是专业工作流的核心思想。
+
+为了让这个想法更具操作性，我们引入一个中间层，这就是大名鼎鼎的 **Git Flow** 思想的简化版，非常适合单人或小团队。
+
+---
+
+### 推荐给你的工作流程 (Solo Dev 最佳实践)
+
+忘掉一次性把所有分支都合到 `main` 的想法，那样会很混乱。采用下面这个更清晰、更安全的流程：
+
+#### 第 1 步：建立两个“永久”分支
+
+1.  **`main` 分支**：
+    *   **用途**：神圣不可侵犯！它只用来存放**已发布的、稳定的、可随时部署的**版本。
+    *   **规则**：你永远不直接向 `main` 提交代码。唯一的例外是通过发布流程从 `develop` 分支合并。你甚至可以在 GitHub 设置中将 `main` 分支保护起来，禁止直接推送。
+
+2.  **`develop` 分支** (或叫 `dev`)：
+    *   **用途**：这是你的**开发主线**。它集成了所有已完成的功能，代表了下一个版本“正在开发中”的最新状态。这个分支应该是大部分时间里功能最全，但可能不完全稳定的地方。
+    *   **操作**：`git checkout -b develop` 从 `main` 创建一次，然后 `git push -u origin develop` 推送到远程。之后它就一直存在了。
+
+#### 第 2 步：针对每个任务的工作流
+
+现在，假设你要开始做项目板上的任务 `Feature: CLI command schema – command skeleton`。
+
+1.  **准备**：确保你的本地 `develop` 分支是最新状态。
+    ```bash
+    git checkout develop
+    git pull origin develop
+    ```
+
+2.  **创建功能分支**：从 `develop` 分支上创建一个新的、描述清晰的功能分支。
+    ```bash
+    # 格式：类型/任务简称
+    git checkout -b feature/cli-skeleton
+    ```
+
+3.  **编码**：在这个 `feature/cli-skeleton` 分支上安心地写代码、修改、测试。频繁地提交 (commit) 是个好习惯。
+    ```bash
+    git add .
+    git commit -m "feat: add basic structure for schema command"
+    # ...继续工作和提交...
+    ```
+
+4.  **推送分支**：当你觉得这个功能基本完成，需要评审或备份时，就把它推送到远程 GitHub 仓库。
+    ```bash
+    git push -u origin feature/cli-skeleton
+    ```
+
+5.  **创建 PR**：现在去你的 GitHub 仓库页面。GitHub 会自动检测到你推送了一个新分支，并提示你创建一个 Pull Request。
+    *   **源分支 (from)**：`feature/cli-skeleton`
+    *   **目标分支 (to)**：`develop`  <-- **注意：是合并到 `develop`，不是 `main`！**
+    *   在 PR 的描述里，链接到你正在做的那个 Issue (例如，写上 `Resolves #2`，#2 是 Issue 的编号)。
+
+6.  **自我审查与合并**：
+    *   即使只有你一个人，也要**认真审查你自己的 PR**。在 “Files changed” 标签页里检查你的每一行改动，这是发现低级错误和改进代码的绝佳机会。
+    *   确认无误后，点击 “Merge Pull Request”。
+    *   合并后，可以选择删除远程的 `feature/cli-skeleton` 分支，因为它已经完成了使命。
+
+7.  **循环**：对于项目板上的下一个任务，重复步骤 1-6。
+
+#### 第 3 步：发布版本 (Release)
+
+当你完成了里程碑 (`v0.1.0`) 中的所有任务，并且它们都已经被合并到了 `develop` 分支，你的 `develop` 分支现在就包含了所有新功能，并且经过了测试。
+
+这时，就到了**发布**的时刻！
+
+1.  **创建发布 PR**：创建一个**新的 PR**，这次是从 `develop` 合并到 `main`。
+    *   **源分支 (from)**：`develop`
+    *   **目标分支 (to)**：`main`
+    *   这个 PR 的标题可以命名为 `Release v0.1.0`。
+
+2.  **最终检查**：这个 PR 包含了你这段时间以来的所有心血。做最后的检查。
+
+3.  **合并到 `main`**：合并这个 PR。现在，你的 `main` 分支就更新到了最新的 `v0.1.0` 版本。
+
+4.  **打上标签 (Tag)**：在 `main` 分支上创建一个 Git 标签，以标记这个发布点。
+    ```bash
+    git checkout main
+    git pull origin main
+    git tag -a v0.1.0 -m "Release version 0.1.0"
+    git push origin v0.1.0
+    ```
+
+### 总结与可视化
+
+你的分支结构看起来会是这样：
+
+```
+main      (稳定版)  <--- [仅在发布时合并] --- develop (开发版)
+                                                ^
+                                                | [PR合并]
+                                                |
+feature/task-A --- (完成) --------------------/
+                                                ^
+                                                | [PR合并]
+                                                |
+feature/task-B --- (完成) --------------------/
+```
+
+这个流程的好处是：
+*   **安全**：`main` 分支永远是干净的。
+*   **清晰**：每个功能都在独立的分支开发，互不干扰。
+*   **有序**：所有完成的功能汇集到 `develop`，便于集成测试和版本准备。
+*   **可追溯**：每个版本发布都有一个清晰的从 `develop`到 `main` 的 PR 和一个 Git 标签。
+
+---
+
+你应该遵循 **一个任务 = 一个分支 = 一个 PR** 的原则。我们来深入探讨一下为什么这是最佳实践，以及在什么罕见情况下可以破例。
+
+---
+
+### 为什么应该避免多个任务合并到一个 PR？
+
+想象一下，几个月后你发现一个 Bug，你想知道是哪次代码更改引入的。你使用 `git blame` 或查看提交历史来追溯。
+
+**场景A (糟糕的方式：多个任务一个 PR)**
+你找到了引入 Bug 的那次提交，它的提交信息是：“*feat: Add CLI skeleton, validate rules, and setup output formatting*”。
+现在你面临几个问题：
+1.  **审查困难**：这个 PR 可能包含了 20 个文件的修改，上千行代码的变动。要从中找出具体是哪个功能的哪一部分引入了问题，就像大海捞针。即使是在提交前自我审查，也很容易因为信息过载而忽略细节。
+2.  **回滚困难 (High Risk)**：假设你发现“规则验证”部分有严重问题，需要立即撤销。但因为它和“CLI 骨架”捆绑在同一个 PR 里，你无法轻易地只回滚那一部分。如果你强行回滚整个 PR，你会连同已经写好的、没有问题的“CLI 骨架”代码也一起弄丢了。**这就把不同功能的风险捆绑在了一起。**
+3.  **历史混乱**：你的项目历史不再是“一步一个脚印”的故事，而是一堆混杂在一起的大杂烩。这使得理解项目演进过程变得异常困难。
+4.  **工作流阻塞**：如果这三个任务在一个 PR 里，而其中一个任务的实现方式你还不满意，需要重构，那么另外两个已经完成的任务也无法合并，只能一起等待。
+
+**场景B (最佳实践：一个任务一个 PR)**
+你找到了引入 Bug 的那次提交，它的信息是：“*feat: Implement rules file validation*”。
+1.  **审查清晰**：这个 PR 只包含与“规则文件验证”相关的 3-4 个文件改动，总共几十行代码。你可以非常专注地审查这些改动，快速定位问题。
+2.  **回滚简单 (Low Risk)**：发现问题后，你可以非常安全地回滚这一个 PR (`git revert <commit_hash>`)，它只会精确地撤销“规则文件验证”的更改，完全不影响项目中其他已经合并的功能。
+3.  **历史清晰**：你的 `develop` 分支历史会像这样：
+    *   `Merge pull request #3 from feature/output-formatting`
+    *   `Merge pull request #2 from feature/rules-validation`
+    *   `Merge pull request #1 from feature/cli-skeleton`
+    这是一部清晰的编年史，任何人（包括未来的你）都能轻松读懂。
+
+---
+
+### 把 PR 当成写作中的“段落”
+
+你可以把每一次 PR 想象成学术写作中的一个**段落**。一个好的段落只阐述一个核心观点。如果你试图在一个段落里同时讨论三个不相关的话题，读者会感到困惑。
+
+同样，一个好的 PR 只实现一个独立的、完整的功能点（一个任务）。这让你的代码库逻辑清晰、易于维护、并且非常安全。
+
+---
+
+### 有没有例外情况？
+
+是的，在一些非常罕见的情况下，可以把极其相关的微小任务合并。
+
+**可以接受的例子：**
+*   **任务1**：在配置文件中添加一个新的配置项 `timeout`。
+*   **任务2**：在 `README.md` 文档中说明这个 `timeout` 配置项的用法。
+
+这两个任务是**强耦合**和**逻辑原子**的。如果只做了任务1而不做任务2，功能就不完整；反之亦然。在这种情况下，在一个名为 `feature/add-timeout-config` 的分支上同时完成这两件事，然后创建一个 PR，是完全合理的。这个 PR 的核心目的依然是单一的：“添加超时配置功能”。
+
+**要避免的例子 (根据你的项目)：**
+你绝对不应该把下面这三个任务放在同一个 PR 里：
+*   `Feature: CLI command schema – command skeleton` (这是结构层)
+*   `Task: Rules file validation (minimal, no jsonschema)` (这是逻辑层)
+*   `Feature: Core – add Schema rule type` (这是核心引擎层)
+
+这三个任务虽然相关，但它们属于软件的不同层面，应该被独立开发、独立测试、独立合并。
+
+### 给你的具体操作建议
+
+1.  **回到你的项目板**，从 `Todo` 列表里只选**一个**任务开始。比如 `Feature: CLI command schema – command skeleton`。
+2.  **创建分支**：`git checkout -b feature/cli-skeleton`。
+3.  **完成工作**：在这个分支上完成该任务的所有代码。
+4.  **创建 PR**：创建一个从 `feature/cli-skeleton` 到 `develop` 的 PR，并链接到对应的 Issue。
+5.  **审查与合并**：自我审查，然后合并它。
+6.  **重复**：回到第一步，去 `develop` 分支上拉取最新代码 (`git checkout develop && git pull`)，然后为你的**下一个任务**创建一个**新的分支**。
+
+这种节奏和纪律，即使对于单人项目，也会在长期带来巨大的回报。你的“未来队友”（也就是几个月后的你）会非常感谢你今天养成的这个好习惯。
