+++ notes/ROADMAP.md
@@ -1,0 +1,53 @@
+# ValidateLite 项目路线图 (Roadmap)
+
+本文档旨在规划 `validatelite` 项目的中长期发展方向，它将随着社区反馈和技术演进进行迭代。
+
+---
+
+### Phase 1: 基础架构增强 (v0.5.0)
+
+此阶段的核心是现代化CLI接口，并为核心功能的多表支持奠定基础。
+
+*   **CLI 接口标准化**
+    *   **目标**: 弃用位置参数 `<data_source>`，全面转向更明确的 `--conn` 和 `--table` 选项，提升命令的清晰度和可扩展性。
+    *   **涉及命令**: `check`, `schema`。
+
+*   **`schema` 命令多表支持**
+    *   **目标**: 允许 `schema` 命令通过单个规则文件验证数据源中的多个表。
+    *   **实现**: 引入以表名为键的JSON文件新格式。
+
+---
+
+### Phase 2: 深度验证能力 (v0.6.0)
+
+此阶段将深化 `schema` 命令的验证能力，使其能够覆盖更精细、更智能的验证场景。
+
+*   **`schema` 验证: 字段长度与精度**
+    *   **目标**: 增加对字符串长度、数值类型精度和小数位数的验证。
+    *   **实现**: 在规则文件的字段定义中增加 `length`, `precision`, `scale` 等属性。
+
+*   **`schema` 验证: "软"类型检查**
+    *   **目标**: 验证某列的值是否可以被安全地转换为预期的类型。
+    *   **实现**: 引入 `expected_type` 规则，例如，检查一个`varchar`列的值是否都是有效的`integer`或`date`格式。
+
+---
+
+### Phase 3: 命令套件扩展 (v1.0.0+)
+
+此阶段标志着 `validatelite` 从一个单一工具向一个多功能的数据质量平台演进，我们将引入为特定业务场景设计的全新独立命令。
+
+*   **新命令: `vlite diff`**
+    *   **定位**: 数据稽核与和解工具。
+    *   **场景**: 对比两个数据集（可跨库）的差异，包括行记录差异、字段值差异和聚合值差异。适用于数据迁移验证、ETL任务校验等。
+
+*   **新命令: `vlite profile`**
+    *   **定位**: 数据剖析与探索工具。
+    *   **场景**: 对数据进行快速的统计分析（如空值率、基数、最值、分布等），在数据接入或分析前，快速评估其“健康状况”，并能生成可视化报告。
+
+---
+
+### 远期设想
+
+*   **数据血缘 (`vlite lineage`)**: 追踪数据字段的来源与流向。
+*   **与数据目录的集成**: 与 `OpenMetadata`, `DataHub` 等工具联动。
+*   **Web UI**: 提供一个图形化界面，用于定义规则、管理任务和查看报告。
