+++ notes/近期计划的需求.md
@@ -1,0 +1,80 @@
+# 近期计划的需求 (v0.5.0)
+
+本文档详细定义了 `validatelite` 下一阶段 (v0.5.0) 的核心开发需求，旨在为支持更复杂的验证场景和未来的功能扩展奠定坚实的基础。
+
+---
+
+### 总览
+
+核心目标是**改进CLI的参数设计**并**增强`schema`命令的功能**。
+
+我们将从当前使用**位置参数 `<data_source>`** 的模式，转向使用更清晰、更灵活的**选项参数（`--conn`, `--table`）**模式。这将是支持多表验证的关键一步。
+
+---
+
+### 需求详情
+
+#### 1. `check` 命令接口标准化
+
+*   **目标**: 使 `check` 命令的接口更明确，将数据源连接信息和表名分离。
+*   **现状 (As-Is)**: `vlite check <connection_and_table_string> --rules <file.json>`
+*   **未来 (To-Be)**: `vlite check --conn <connection_string> --table <table_name> --rules <file.json>`
+*   **验收标准**:
+    1.  移除对位置参数 `<data_source>` 的依赖。
+    2.  引入两个新的、必需的选项参数：
+        *   `--conn <string>`: 用于指定数据库连接字符串或文件路径。
+        *   `--table <string>`: 用于指定要验证的表名或文件名。
+    3.  旧的 `vlite check <data_source>` 格式应被标记为“已弃用”，并给出明确的提示信息，引导用户使用新格式。
+    4.  更新 `vlite check --help` 帮助文档，反映新的参数设计。
+
+#### 2. `schema` 命令功能增强：支持多表验证
+
+*   **目标**: 使 `schema` 命令能够使用单个规则文件，一次性验证指定数据源中的多个表结构。
+*   **现状 (As-Is)**: `vlite schema <connection_and_table_string> --rules <single_table_schema.json>`
+*   **未来 (To-Be)**: `vlite schema --conn <connection_string> --rules <multi_table_schema.json>`
+*   **验收标准**:
+    1.  **CLI接口变更**: 
+        *   与 `check` 命令类似，移除对位置参数 `<data_source>` 的依赖。
+        *   引入必需的 `--conn <string>` 选项参数。
+        *   **注意**: `schema` 命令在多表模式下**不**需要 `--table` 参数，因为所有待验证的表都将在规则文件中定义。
+    2.  **重新定义 `--rules` 文件结构**:
+        *   为支持多表，引入新的JSON结构：一个以**表名为键**，以**该表的schema定义为值**的顶层对象。
+        *   **示例 (multi_table_schema.json)**:
+            ```json
+            {
+              "users": {
+                "rules": [
+                  { "field": "id", "type": "integer", "required": true },
+                  { "field": "age", "type": "integer", "min": 0, "max": 120 },
+                  { "field": "gender", "type": "string", "enum": ["M", "F"] },
+                  { "field": "email", "type": "string", "required": true },
+                  { "field": "created_at", "type": "datetime" }
+                ]
+              },
+              "products": {
+                "rules": [
+                  { "field": "product_id", "type": "integer" },
+                  { "field": "price", "type": "float" }
+                ],
+                "strict_mode": false
+              }
+            }
+            ```
+    3.  **更新 `schema` 命令执行逻辑**:
+        *   程序加载 `--rules` 文件后，遍历JSON对象的所有顶级键（`users`, `products`等）。
+        *   对于每个键（表名），从 `--conn` 指定的数据源中获取该表的实际schema信息。
+        *   将实际schema与规则文件中定义的期望schema进行比对。
+    4.  **优化输出信息**: 验证报告必须清晰地分组，指明每个表的验证结果。
+        *   **示例输出**:
+            ```
+            Schema validation results for connection: mysql://...
+
+            [Table: users]
+            Column 'id': PASSED
+            Column 'email': PASSED
+
+            [Table: products]
+            Column 'product_id': PASSED
+            Column 'price': FAILED - Expected type 'float', found 'decimal'.
+            ```
+    5.  更新 `vlite schema --help` 帮助文档。
